refactor: make socks implementation to be in single file
