+++ lib/exceptions.ts
@@ -1,9 +1,3 @@
-export class TorException extends Error {
-  constructor(message: string) {
-    super(`[SOCKS5]: ${message}`);
-  }
-}
-
 export class TorHttpException extends Error {
   constructor(message: string) {
     super(`[HTTP]: ${message}`);
